refactor(tor): keep the transport out of the domain
