+++ packages/bull_tor/lib/tor.dart
@@ -7,13 +7,13 @@
 library;
 
 export 'src/data/tor_http_client_factory.dart';
+export 'src/data/tor_connection_failure_recorder.dart';
 export 'src/domain/entities/tor_connection_state.dart';
 export 'src/domain/entities/tor_proxy_endpoint.dart';
 export 'src/domain/entities/tor_route.dart';
 export 'src/domain/entities/tor_session.dart';
 export 'src/domain/entities/tor_transport.dart';
 export 'src/domain/onion_connection_failure.dart';
-export 'src/domain/tor_connection_failure_recorder.dart';
 export 'src/domain/tor_route_pool.dart';
 export 'src/domain/tor_failure.dart';
 export 'src/domain/usecases/ensure_tor_ready_usecase.dart';
@@ -21,5 +21,3 @@ export 'src/domain/usecases/retry_tor_connection_usecase.dart';
 export 'src/domain/usecases/verify_external_tor_usecase.dart';
 export 'src/domain/usecases/watch_tor_connection_usecase.dart';
 export 'src/tor_controller.dart';
-export 'package:socks5_proxy/enums.dart';
-export 'package:socks5_proxy/exceptions.dart';
